Move Spotify Manager to separate class, code clean up
